pyproject.toml
[tool.poetry]
name = "arcade_local_file_management"
version = "0.1.0"
description = "LLM tools for tasks relating to folder and file management that are stored locally"
authors = ["Eric Gustin <eric@arcade-ai.com>"]

[tool.poetry.dependencies]
python = "^3.10"
arcade-ai = "^0.1.*"

[tool.poetry.dev-dependencies]
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"